# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CampaignCorrectAnswer = typeof CampaignCorrectAnswer[keyof typeof CampaignCorrectAnswer];


export const CampaignCorrectAnswer = {
  legitimate: 'legitimate',
  phishing: 'phishing',
} as const;

export type CampaignDifficulty = typeof CampaignDifficulty[keyof typeof CampaignDifficulty];


export const CampaignDifficulty = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
} as const;

export interface Campaign {
  id: string;
  title: string;
  description: string;
  sender: string;
  subject: string;
  content: string;
  link: string;
  suspiciousClues: string[];
  correctAnswer: CampaignCorrectAnswer;
  explanation: string;
  category: string;
  difficulty?: CampaignDifficulty;
  createdAt?: string;
}

export type CampaignInputCorrectAnswer = typeof CampaignInputCorrectAnswer[keyof typeof CampaignInputCorrectAnswer];


export const CampaignInputCorrectAnswer = {
  legitimate: 'legitimate',
  phishing: 'phishing',
} as const;

export type CampaignInputDifficulty = typeof CampaignInputDifficulty[keyof typeof CampaignInputDifficulty];


export const CampaignInputDifficulty = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
} as const;

export interface CampaignInput {
  /**
     * @minLength 3
     * @maxLength 120
     */
  title: string;
  /**
     * @minLength 10
     * @maxLength 500
     */
  description: string;
  /**
     * @minLength 3
     * @maxLength 160
     */
  sender: string;
  /**
     * @minLength 3
     * @maxLength 160
     */
  subject: string;
  /**
     * @minLength 10
     * @maxLength 3000
     */
  content: string;
  /**
     * @minLength 3
     * @maxLength 300
     */
  link: string;
  /**
     * @minItems 1
     * @maxItems 8
     * @items.minLength 3
     * @items.maxLength 200
     */
  suspiciousClues: string[];
  correctAnswer: CampaignInputCorrectAnswer;
  /**
     * @minLength 10
     * @maxLength 1000
     */
  explanation: string;
  /**
     * @minLength 3
     * @maxLength 80
     */
  category: string;
  difficulty?: CampaignInputDifficulty;
}

export type CampaignUpdate = CampaignInput;

export type CampaignAttemptInputAnswer = typeof CampaignAttemptInputAnswer[keyof typeof CampaignAttemptInputAnswer];


export const CampaignAttemptInputAnswer = {
  legitimate: 'legitimate',
  phishing: 'phishing',
} as const;

export interface CampaignAttemptInput {
  answer: CampaignAttemptInputAnswer;
}

export interface CampaignAttemptResult {
  correct: boolean;
  answer: string;
  correctAnswer: string;
  explanation: string;
}

export interface QuizQuestion {
  id: string;
  question: string;
  options: string[];
  category: string;
  explanation?: string;
}

export interface QuizQuestionInput {
  /**
     * @minLength 10
     * @maxLength 500
     */
  question: string;
  /**
     * @minItems 2
     * @maxItems 5
     * @items.minLength 1
     * @items.maxLength 240
     */
  options: string[];
  /**
     * @minimum 0
     * @maximum 4
     */
  correctAnswer: number;
  /**
     * @minLength 3
     * @maxLength 80
     */
  category: string;
  /**
     * @minLength 10
     * @maxLength 500
     */
  explanation: string;
}

export type QuizQuestionUpdate = QuizQuestionInput;

export type QuizSubmissionAnswersItem = {
  questionId: string;
  /**
     * @minimum 0
     * @maximum 4
     */
  answerIndex: number;
};

export interface QuizSubmission {
  answers: QuizSubmissionAnswersItem[];
}

export type QuizResultLevel = typeof QuizResultLevel[keyof typeof QuizResultLevel];


export const QuizResultLevel = {
  Faible: 'Faible',
  Moyen: 'Moyen',
  Bon: 'Bon',
  Très_bon: 'Très bon',
} as const;

export interface QuizResult {
  /**
     * @minimum 0
     * @maximum 100
     */
  score: number;
  /** @minimum 0 */
  correctAnswers: number;
  /** @minimum 0 */
  totalQuestions: number;
  level: QuizResultLevel;
  recommendation: string;
  completedAt: string;
}

export type RecommendationPriority = typeof RecommendationPriority[keyof typeof RecommendationPriority];


export const RecommendationPriority = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface Recommendation {
  id: string;
  title: string;
  message: string;
  focus: string;
  priority: RecommendationPriority;
}

export interface Dashboard {
  campaignsCompleted: number;
  quizzesCompleted: number;
  latestScore: number;
  level: string;
  streak: number;
  nextAction: string;
}

export type AdminStatsLevelsItem = {
  label: string;
  value: number;
};

export interface AdminStats {
  users: number;
  campaigns: number;
  quizzesCompleted: number;
  averageScore: number;
  successRate: number;
  levels: AdminStatsLevelsItem[];
}