# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminStatsLevelsItem } from './adminStatsLevelsItem';

export interface AdminStats {
  users: number;
  campaigns: number;
  quizzesCompleted: number;
  averageScore: number;
  successRate: number;
  levels: AdminStatsLevelsItem[];
}