# lib/api-zod/src/generated/types/campaignInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CampaignInputCorrectAnswer } from './campaignInputCorrectAnswer';
import type { CampaignInputDifficulty } from './campaignInputDifficulty';

export interface CampaignInput {
  /**
     * @minLength 3
     * @maxLength 120
     */
  title: string;
  /**
     * @minLength 10
     * @maxLength 500
     */
  description: string;
  /**
     * @minLength 3
     * @maxLength 160
     */
  sender: string;
  /**
     * @minLength 3
     * @maxLength 160
     */
  subject: string;
  /**
     * @minLength 10
     * @maxLength 3000
     */
  content: string;
  /**
     * @minLength 3
     * @maxLength 300
     */
  link: string;
  /**
     * @minItems 1
     * @maxItems 8
     * @items.minLength 3
     * @items.maxLength 200
     */
  suspiciousClues: string[];
  correctAnswer: CampaignInputCorrectAnswer;
  /**
     * @minLength 10
     * @maxLength 1000
     */
  explanation: string;
  /**
     * @minLength 3
     * @maxLength 80
     */
  category: string;
  difficulty?: CampaignInputDifficulty;
}