# lib/api-zod/src/generated/types/quizSubmission.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QuizSubmissionAnswersItem } from './quizSubmissionAnswersItem';

export interface QuizSubmission {
  answers: QuizSubmissionAnswersItem[];
}